fputcsv() bug fixed, current date added to graph.
The graph displayed on the stats page looks a lot more like it should in the final version of the webpage. The next commit should have code written such that a row of data is appended to the .csv file each day. This also fixed errors relating to the incorrect commenting of PHP code that was not spotted before.
This still needs to be integrated with real-world data but given how the code is written as of this commit, this should not take too long.